fix: memorize, change directory, refactoring
