+++ components/organisms/AddButton/AddButton.tsx
@@ -1,0 +1,54 @@
+import React, { useRef, useState } from 'react';
+import { NextPage } from 'next';
+import {
+  Popover, PopoverArrow, PopoverBody, PopoverContent, PopoverTrigger, useOutsideClick,
+} from '@chakra-ui/react';
+
+import { Button } from 'components/Button';
+import { AddForm } from './AddForm';
+
+type Props = {
+  nextId: number;
+};
+
+export const AddButton: NextPage<Props> = ({ nextId }) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const initialFocusRef = useRef<HTMLInputElement>(null);
+
+  const popoverRef = useRef<HTMLElement>(null);
+  useOutsideClick({
+    ref: popoverRef,
+    handler: () => setIsOpen(false),
+  })
+
+  const handleClickAdd = () => {
+    setIsOpen(true);
+  };
+
+  const handleSubmit = async () => {
+    setIsOpen(false);
+  };
+
+  return (
+    <Popover
+      isOpen={isOpen}
+      initialFocusRef={initialFocusRef}
+    >
+      <PopoverTrigger>
+        <Button onClick={handleClickAdd}>
+          Add
+        </Button>
+      </PopoverTrigger>
+      <PopoverContent ref={popoverRef} color="gray.700">
+        <PopoverArrow />
+        <PopoverBody>
+          <AddForm
+            id={nextId}
+            firstInputRef={initialFocusRef}
+            onSubmit={handleSubmit}
+          />
+        </PopoverBody>
+      </PopoverContent>
+    </Popover>
+  );
+};
